chore: refactor handler
